Just some cleanup
