+++ src/workflows/airqo_etl_utils/config.py
@@ -626,7 +626,7 @@ class Config:
     MONGO_URI = os.getenv("MONGO_URI")
     MONGO_DATABASE_NAME = os.getenv("MONGO_DATABASE_NAME", "airqo_db")
     ENVIRONMENT = os.getenv("ENVIRONMENT")
-    CALIBRATEBY = os.getenv("CALIBRATEBY")
+    CALIBRATEBY = os.getenv("CALIBRATEBY", "country")
 
     # Twitter bot
     TWITTER_BOT_API_KEY = os.getenv("TWITTER_BOT_API_KEY")
